Move MathTools class into a separate resource file.
